requirements.txt
Django
djangorestframework
bcrypt
django-cors-headers
django-allauth
Pillow